Fix display.Name item tag in configmenu.yml interfering with I18nUtil.tr in recent version of Paper
It is needed to set the display name of an item in Json format to avoid implicit interpretation of the value of the tag. Here, the 2 items was named "'{'" and "'}'", but the resent version of Paper interpreted it as "{" and "}" without the single quotes, messing with the formatter of I18nUtil.tr() method.