Fix wither spawning with disabled spawn-limits (#66)
* Fix wither spawning with disabled spawn-limits

Resolves #60

Decouples wither-origin tracking and spawn limit listeners. The issue was that withers were tagged with their origin island in the creature spawn listeners for of SpawnEvents, which is not registered if spawn limits are disabled. I solved this by moving the wither tagging logic into a separate listener that is always active. Untagged Withers are despawned by the GriefEvents listener.

I also migrated the tagging to PDC to make the tagging consistent without relying on name tagging that is prone to errors.

* Cleanup SpawnEvents