Fix item enchant not being actually applied to the menu items in the challenge menu (#73)
Completed challenges are usually shown with an enchant glowing effect when they are completed at least once. This may be useful for the player, but the "enchanted" status of the item is also used in the ChallengeLogic to determine if a challenge with `offset: -1` is displayed or not, in place of the completed one.

New MC version, with the new data component system, does not allow an enchant level of 0 (even with unsafe enchant level), an is silently ignored.

This commit changes the enchant level to 1, and adds a flag to the item meta to hide the enchant information from the tooltip.